requirements.txt
# PDF Processing Libraries
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
pypdfium2>=4.30.0
pdf2image>=1.17.0

# OCR Libraries
pytesseract>=0.3.10
# Note: Tesseract must be installed separately:
# macOS: brew install tesseract tesseract-lang
# Ubuntu: sudo apt-get install tesseract-ocr tesseract-ocr-kor

# Image Processing
Pillow>=10.3.0
opencv-python>=4.9.0

# AI/LLM Integration
anthropic>=0.34.0
openai>=1.40.0

# Utilities
click>=8.1.0
tqdm>=4.66.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
asyncio>=3.4.3

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Logging and Monitoring
loguru>=0.7.0
rich>=13.0.0